Finished the class but still need to test.
NOTE: Fixed the infinite loop in bubbledown, not updaint the left and right variable.
Fixed mapSwap, forgot to insert new indexes and was trying to erase old index by the T value insted of index.
Fixed the remove() function with alwaays returning false, issue was caused by mapAdd and mapSwap not functioning correctly.  Function was also creating a new set and checking for set emptiness.

Fixed: issue with array based construnctor not creating a max heap in for loop.